% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_responses.R
\name{find_closest_token}
\alias{find_closest_token}
\title{find closest token for mispoelled token}
\usage{
find_closest_token(
  misspelled_df,
  correct_df,
  accept_threshold = 0.8,
  close_match_treshold = 0.8
)
}
\arguments{
\item{misspelled_df}{a spacyr object with the tokens without a vector (incorrectly spelled)}

\item{correct_df}{a spacyr object with the tokens with a vector (correctly spelled)}

\item{accept_threshold}{the minimum similarity score to accept a match}

\item{close_match_treshold}{the minimum similarity score to accept a proposed match}
}
\value{
a preprocessed character vector
}
\description{
find closest token for mispoelled token
}
