% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_responses.R
\name{load_stopwords}
\alias{load_stopwords}
\title{loading stopwords from the text file in the package}
\usage{
load_stopwords()
}
\value{
a character vector with stopwords
}
\description{
loading stopwords from the text file in the package
}
