% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlirp-package.R
\docType{package}
\name{nlirp-package}
\alias{nlirp}
\alias{nlirp-package}
\title{nlirp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
