% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_responses.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Preprocess a character vector}
\usage{
preprocess_text(x, preprocess_params = list())
}
\arguments{
\item{x}{a character vector}

\item{preprocess_params}{a list of parameters to be passed to the function, the default parameters are; lowercase = TRUE, rm_special_characters = TRUE, rm_qti = TRUE, rm_digits = FALSE, rm_stopwords = TRUE}
}
\value{
a preprocessed character vector
}
\description{
Preprocess a character vector
}
