% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_responses.R
\name{prop_score_sim}
\alias{prop_score_sim}
\title{calculate probabilities for possible scores based on similarity}
\usage{
prop_score_sim(
  data,
  possible_item_scores,
  min_sim_weight = 0.8,
  n_highest_sim = 5,
  start_weight = 0.5,
  min_obs_accept = 5,
  min_certainty_accept = 0.75
)
}
\arguments{
\item{data}{a data.frame with the response_data after preprocesssing, spellcheck and vectorization with nlp}

\item{possible_item_scores}{a vector with the possible scores for the responses}

\item{min_sim_weight}{the minimum similarity weight to accept a similarity}

\item{n_highest_sim}{the number of highest similarities above the min_sim_weight to accept}

\item{start_weight}{the weight to start with for the similarity on the min_sim_weight}

\item{min_obs_accept}{the minimum number of observations to accept a score}

\item{min_certainty_accept}{the minimum certainty to accept a score}
}
\value{
a data.frame with the expected scores for each response
}
\description{
calculate probabilities for possible scores based on similarity
}
