% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_responses.R
\name{score_responses}
\alias{score_responses}
\title{Get expected scores based on semantic similarity using spacy nlp nl_core_news_lg model}
\usage{
score_responses(
  response_data,
  possible_item_scores = NULL,
  preprocessing = TRUE,
  spellcheck = TRUE,
  pos_keep = c("NOUN", "PROPN", "VERB", "ADJ", "ADV"),
  min_sim_weight = 0.8,
  n_highest_sim = 5,
  start_weight = 0.5,
  min_obs_accept = 5,
  min_certainty_accept = 0.75,
  additonal_parameters = list(preprocess_params = list(), spellcheck_params = list())
)
}
\arguments{
\item{response_data}{a data.frame or tibble with 3 columns (response_id, response, score) where response_id is the unique identifier for each response, response is the text of the response, and score is the score of the response given by a human rater (the response can be NA)}

\item{possible_item_scores}{a vector with the possible scores for the responses}

\item{preprocessing}{logical; if TRUE preprocess the text before calculating the similarity}

\item{spellcheck}{logical; if TRUE apply spellcheck the text before calculating the similarity}

\item{pos_keep}{a character vector with the parts of speech to keep}

\item{min_sim_weight}{a numeric value between 0 and 1; the minimum similarity weight to accept a similarity}

\item{n_highest_sim}{an integer value; the number of highest similarities above the min_sim_weight to accept}

\item{start_weight}{a numeric value between 0 and 1; the weight to start with for the similarity on the min_sim_weight}

\item{min_obs_accept}{an integer value; the minimum number of observations to accept a score}

\item{min_certainty_accept}{a numeric value between 0 and 1; the minimum certainty to accept a score}

\item{additonal_parameters}{a list of parameters to be passed to the function}
}
\value{
The original data.frame from data with an added columns
}
\description{
Get expected scores based on semantic similarity using spacy nlp nl_core_news_lg model
}
