% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_responses.R
\name{spellcheck_text}
\alias{spellcheck_text}
\title{spellcheck a character vector}
\usage{
spellcheck_text(spacy_object, spellcheck_params = list())
}
\arguments{
\item{spacy_object}{a character vector}

\item{spellcheck_params}{a list of parameters to be passed to the function, the default parameters are; spellcheck_closeness = .99, pos_to_check = c("NOUN", "PROPN", "VERB", "ADJ", "ADV"), pos_to_use = c("NOUN", "PROPN", "VERB", "ADJ", "ADV")}
}
\value{
a preprocessed character vector
}
\description{
spellcheck a character vector
}
